[godot3] Another huge round of changes
* All (all!) scenes have been reworked so that they are now resolution
independent. Scaling is not enabled in the project settings, so font
size does not scale, and all GUI scenes stretch to fill all the space
* As part of the last point, the game scene has been reworked to utilise
a camera that automatically adjusts zoom so that the whole game level is
centered on the screen and as large as possible (keeping the 16:9 aspect
ratio).
    * This also opens up the opportunity for camera effects (screenshake
    etc), woohoo!
    * To mask shitty edges of the levels, added "overlays" in the
    horizontal and vertical directions, which are just panels extending
    out to +/- 100000 in those directions. Bit of a hack, but seems to
    work. These panels are white during gameplay, and darken to black
    when paused/controller disconnect/end game scenes are shown.
* Main font size increased to 64. Because we are no longer scaling the
game, the font looks crisp at all resolutions!